Develop

+++ include/logger.h
@@ -55,7 +55,7 @@ namespace exafmm {
     double get_time() {
       struct timeval tv;                                        // Time value
       gettimeofday(&tv, NULL);                                  // Get time of day in seconds and microseconds
-      return double(tv.tv_sec+tv.tv_usec*1e-6);                 // Combine seconds and microseconds and return
+      return double(tv.tv_sec)+double(tv.tv_usec)*1e-6;         // Combine seconds and microseconds and return
     }
 
     //! Cycle counter
